standalone_with_docker/Consumers/requirements.txt
pulsar-client==2.7.1
